+++ Back/Events/src/main/java/com/example/supplier/Service/RatingService.java
@@ -8,12 +8,12 @@ import java.util.List;
 public interface RatingService {
 
 public void save(Rating rating);
-public void remove(long event,long userId);
+public void remove(int event,int userId);
     public void remove(Rating rating);
-    public List<Rating> findUserRatings(long user);
-    public List<Rating> findEventRatings(long event);
+    public List<Rating> findUserRatings(int user);
+    public List<Rating> findEventRatings(int event);
     public List<Rating> findAllRatings();
 
     public void updateRating(Rating rating);
-    public void updateRatingForEvent(long event);
+    public void updateRatingForEvent(int event);
 }
